requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
streamlit==1.38.0
pandas==2.2.2
plotly==5.24.1
folium==0.17.0
requests==2.32.3
python-dotenv==1.0.1
sqlalchemy==2.0.36
Pillow==10.4.0
scikit-learn==1.5.2
reportlab==4.2.5
exifread==3.0.0
streamlit-folium==0.17.0